Delete flaky test, refactor LoggerBuilder to be idempotent
